requirements.txt
streamlit>=1.18.0
ccxt>=3.0.0
pandas>=1.5.0
numpy>=1.23.0
pytz>=2022.1
pyyaml>=6.0
plotly>=5.13.0
matplotlib>=3.7.0
xlsxwriter>=3.0.0
openpyxl>=3.0.0
requests>=2.28.0 